+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All features and notable changes to this project will be documented in this file
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [0.0.4] - 2023-03-02
+
+### Added
+- handling interrupt signal
+- using pool to listen for keyboard and socket input
+
 ## [0.0.3] - 2023-02-29
 
 ### Added